requirements.txt
GitPython
click
docker
giturlparse
pydantic
requests
ruamel.yaml
termcolor
